% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_multimodal.R
\name{build_multimodal}
\alias{build_multimodal}
\title{Build multimodal trips from trajectories}
\usage{
build_multimodal(
  data = NULL,
  spatial_threshold,
  temporal_threshold,
  palmsplus = NULL,
  verbose = TRUE,
  multimodal_fields = NULL,
  trajectory_locations = NULL
)
}
\arguments{
\item{data}{The trajectories object built with \code{palms_calc_trajectories}.}

\item{spatial_threshold}{Spatial threshold in meters}

\item{temporal_threshold}{Temporal threshold in minutes}

\item{palmsplus}{The dataset build by \code{build_hbGIS}}

\item{verbose}{Print progress after each step. Default is \code{TRUE}.}

\item{multimodal_fields}{...}

\item{trajectory_locations}{...}
}
\value{
The input trajectories LINESTRING geometry, collapsed into multimodal trips
}
\description{
Build multimodal trips from trajectories.
}
\details{
Several columns are required in the \code{trajectories} dataset. These
need to be added as trajectory fields:
\itemize{
\item identifier
\item tripnumber
\item mot
\item start
\item end
\item geometry
}
}
