% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_trajectories.R
\name{build_trajectories}
\alias{build_trajectories}
\title{Build trajectories from the hbGIS dataset}
\usage{
build_trajectories(
  data = NULL,
  trajectory_fields = NULL,
  trajectory_locations = NULL
)
}
\arguments{
\item{data}{The dataset build by \code{build_hbGIS}.}

\item{trajectory_fields}{trajectory_fields}

\item{trajectory_locations}{trajectory_locations}
}
\value{
A table of individual trips represented as \code{LINESTRING} geometry.
}
\description{
Build trajectories (trips) from the \code{hbGPS} dataset. This
returns a \code{sf data.frame} with \code{LINESTRING} geometry. Three columns
are returned by default (\code{identifier}, \code{tripnumber}, and \code{geometry}).
}
