% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_whenwhatwhere.R
\name{build_whenwhatwhere}
\alias{build_whenwhatwhere}
\title{build_whenwhatwhere}
\usage{
build_whenwhatwhere(
  data = NULL,
  verbose = TRUE,
  whenwhat_field = NULL,
  loca = NULL,
  participant_basis = NULL
)
}
\arguments{
\item{data}{The hbGPS data obtained using \code{read_palms} from palmplusr.}

\item{verbose}{Print progress to console after each iteration. Default is \code{TRUE}.}

\item{whenwhat_field}{whenwhat_field defined in hbGIS}

\item{loca}{Nested list with location information}

\item{participant_basis}{participant_basis}
}
\description{
Build the whenwhat dataset by adding additional columns to the
hbGPS output data based on whenwhat_field.
}
