% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbGIS.R
\name{hbGIS}
\alias{hbGIS}
\title{palmsplusr_shiny}
\usage{
hbGIS(
  gisdir = "",
  palmsdir = "",
  gislinkfile = "",
  outputdir = "",
  dataset_name = "",
  configfile = "",
  verbose = TRUE
)
}
\arguments{
\item{gisdir}{Path to directory with GIS files}

\item{palmsdir}{Path to hbGPS output directory}

\item{gislinkfile}{Path to participant basis file, which is the file that links all participant identifies with the GIS data}

\item{outputdir}{Path to outputdir location}

\item{dataset_name}{Name of dataset}

\item{configfile}{Configuration file}

\item{verbose}{verbose Boolean}
}
\value{
palms_to_clean_lower object
}
\description{
palmsplusr_shiny
}
