% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_params.R
\name{load_params}
\alias{load_params}
\title{load_params}
\usage{
load_params(file = c())
}
\arguments{
\item{file}{Character to specify location of configuration file}
}
\value{
list of parameters extract from the configuration file
}
\description{
load_params
}
