% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_params.R
\name{update_params}
\alias{update_params}
\title{update_params}
\usage{
update_params(new_params = c(), file = c())
}
\arguments{
\item{new_params}{New parameters}

\item{file}{Character to specify location of original configuration file}
}
\value{
No object returned, function only reads original data, and overwrites parameters and stores it again
}
\description{
update_params
}
