% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_days.R
\name{build_days}
\alias{build_days}
\title{Calculate day-level summaries from the whenwhat dataset}
\usage{
build_days(
  whenwhatwhere = NULL,
  verbose = TRUE,
  where_field = NULL,
  whenwhat_field = NULL,
  loca = NULL
)
}
\arguments{
\item{whenwhatwhere}{The whenwhatwhere data obtained from \code{\link{build_whenwhatwhere}}.}

\item{verbose}{Print progress to console. Default is \code{TRUE}.}

\item{where_field}{...}

\item{whenwhat_field}{...}

\item{loca}{Nested list with location information}
}
\value{
A table summarised by day.
}
\description{
Build a days dataset by summarising \code{whenwhat}
by day and person (\code{identifier}).
}
