% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbGIS.R
\name{hbGIS}
\alias{hbGIS}
\title{hbGIS}
\usage{
hbGIS(
  gisdir = "",
  palmsdir = "",
  gislinkfile = "",
  outputdir = "",
  dataset_name = "",
  configfile = "",
  verbose = TRUE,
  baselocation = NULL,
  groupinglocation = NULL,
  write_shp = NULL,
  split_GIS = NULL,
  sublocationID = NULL
)
}
\arguments{
\item{gisdir}{Path to directory with GIS files}

\item{palmsdir}{Path to hbGPS output directory}

\item{gislinkfile}{Path to participant basis file, which is the file that links all participant identifies with the GIS data}

\item{outputdir}{Path to outputdir location}

\item{dataset_name}{Name of dataset}

\item{configfile}{Configuration file}

\item{verbose}{verbose Boolean}

\item{baselocation}{character, to specify reference location for individuals, e.g. home}

\item{groupinglocation}{character, to specify reference location location for groups, e.g. school}

\item{write_shp}{boolean, to indicate whether shape file should be written}

\item{split_GIS}{boolean, to indicate whether sublocation inside GIS files are to be split}

\item{sublocationID}{character, GIS column name to be used as identifier for sublocations}
}
\value{
palms_to_clean_lower object
}
\description{
hbGIS
}
